+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 * Fixed Photon node crash
 * Added an option for whether the block is `collisionShapeFullBlock`
 * Added machine translated ru_ru lang file (thanks to @серый калчедан)
+* Added fluid auto world io